renamings and consolidations
