+++ create_dataset.py
@@ -15,10 +15,10 @@ import argparse
 import logging
 
 from config.config_logging import setup_logging
-from data_processing.data_inventory import DatasetInventoryMaster
-from data_processing.data_writer import DatasetWriter
-from data_processing.tfr_encoder_decoder import DefaultTFRecordEncoderDecoder
-from pre_processing.image_transformations import resize_jpeg
+from data.inventory import DatasetInventoryMaster
+from data.writer import DatasetWriter
+from data.tfr_encoder_decoder import DefaultTFRecordEncoderDecoder
+from data.image import resize_jpeg
 
 # Configure Logging
 setup_logging()
