+++ data/tfr_encoder_decoder.py
@@ -3,7 +3,7 @@ import logging
 
 import tensorflow as tf
 
-from data_processing.utils import (
+from data.utils import (
         wrap_int64, wrap_bytes, wrap_dict_bytes_list,
         _bytes_feature_list,
         _bytes_feature_list_str,
